requirements.txt
# BP Badminton Dependencies
# IMPORTANT: Keep this file updated when adding/removing packages
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Limiter==3.5.0
Flask-Caching==2.1.0
Flask-Migrate==4.0.7
python-dotenv==1.0.0
Werkzeug==3.0.1
gunicorn==21.2.0
Flask-Compress==1.14
psycopg2-binary==2.9.9
pyopenssl
cryptography
requests==2.31.0
beautifulsoup4==4.12.3
playwright>=1.40.0
browser-cookie3>=0.19.1
bleach>=6.0.0
google-auth>=2.29.0
